refactoring and reorganising the code
