+++ Akavache.SqlServerCompact/Akavache.SqlServerCompact.csproj
@@ -60,6 +60,7 @@
     <Compile Include="Ensure.cs" />
     <Compile Include="Extensions.cs" />
     <Compile Include="Extensions.CacheElement.cs" />
+    <Compile Include="Extensions.SchemaInfo.cs" />
     <Compile Include="SqlServerCompactPersistentBlobCache.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
