+++ Akavache.SqlServerCompact/CacheElement.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Data.Common;
 
 namespace Akavache.SqlServerCompact
 {
@@ -9,5 +10,23 @@ namespace Akavache.SqlServerCompact
         public byte[] Value { get; set; }
         public DateTime Expiration { get; set; }
         public DateTime CreatedAt { get; set; }
+
+        internal static CacheElement FromDataReader(DbDataReader result)
+        {
+            var cacheElement = new CacheElement
+            {
+                Key = result.GetFieldValue<string>(0),
+                Value = result.GetFieldValue<byte[]>(2),
+                CreatedAt = result.GetFieldValue<DateTime>(3),
+                Expiration = result.GetFieldValue<DateTime>(4)
+            };
+
+            var typeName = result.GetValue(1);
+            if (typeName != DBNull.Value)
+            {
+                cacheElement.TypeName = typeName.ToString();
+            }
+            return cacheElement;
+        }
     }
 }
