+++ Akavache.SqlServerCompact/Extensions.cs
@@ -5,19 +5,14 @@ using System.Reactive.Linq;
 
 namespace Akavache.SqlServerCompact
 {
-    public static partial class Extensions
+    internal static partial class Extensions
     {
-        public static IObservable<Unit> CreateSchemaInfoTable(this SqlCeConnection connection)
+        internal static IObservable<Unit> ExecuteAsync(this SqlCeConnection connection, string text)
         {
             return Observable.Return(Unit.Default);
         }
 
-        public static IObservable<Unit> ExecuteAsync(this SqlCeConnection connection, string text)
-        {
-            return Observable.Return(Unit.Default);
-        }
-
-        public static IObservable<T> ExecuteScalarAsync<T>(this SqlCeConnection connection, string text)
+        internal static IObservable<T> ExecuteScalarAsync<T>(this SqlCeConnection connection, string text)
         {
             return Observable.Return(default(T));
         }
